Remove runtime earl dependency
